Move `TwoFactorAuthenticatable#legacy_otp_secret` into upgrading guide
`TwoFactorAuthenticatable#legacy_otp_secret` is only required for a
short period of time while the user completes all phases of the upgrade
so instead of adding it to the gem, we are including the implemenation
in the upgrade guide for teams to copy & paste into their user models.
This will make it easier for teams who have customised the encryption
scheme used to store the OTP secret and avoid having this gem carry a
complex method which it doesn't really need to have in the longer term.